Rename DefaultLogger
